DEL: deduplicate workflow
we've articulate the workflow in detail on TREP-7. Here it was duplicate (and wrong), so we'll have to choose where to put it.
Looking at the actions on the workflow, I think it's more logical for it to be in the TREP template file (where a user can see all the headers)